fix: rework adjacency list graph vertex traversal checking
Previously, a vertex's traversal state was determined with a boolean, while
now a graph is only considered completely traversed once all of the paths
leading to it have been traversed. So while a vertex is always traversed,
it is only 'visited' on its first traversal. However, further testing shows
that bi-direction paths cannot be formed without creating an infinite loop.