fix: rework how ListGraph object manages GraphVertex objects
In order to add to new vertices to a ListGraph from one of its functions,
we must specify a start and end key, both of which must be searched for
in order to connect one GraphVertex to another. This requires the depth
first search on the GraphVertex end to be able to return to the
ListGraph its shared_ptr so that it can be connected to another if a
key is located. Reference paramameters for all functions have therefore
been changed to shared_ptr parameters, and depth first searching now
requires that the shared_ptr to the calling vertex is passed.