feat: implement graphical representation of adjacenecy list graph
This feature includes a wrapper for a ListGraph, the GUIGraph, comprising
the ListGraph field which contains GUIVertex vertex items. The
GUIVertex items subsequently become the data associated with
the GraphVertex objects that the ListGraph uses as pointers to one another.
Future updates will included implementing functions within GUIGraph that
can alter the GUIVertex graphical representation of the vertices, and for
traversals to update the appearance, positioning, etc., of these vertices