fix: rework depth-first traversal
A depth-first traverse now involves marking each vertex as traversed via
boolean and pushing each traversed vertex to a stack, with which we can
pop from after the traversal to flag each vertex as not traversed. This
method enables us now the ability to perform some sort of action on all
the vertices as the traversal is reversed. However, this has not yet been
implemented at this stage.