feat: add lambda parameter overload for ListGraph vertex adding
This new overload allows the caller to perform initialization on the vertex data reference, as adding new vertices with the data passed as arguments for the start and end key creates pointers to graph vertices with copies of the data. This disallowed initialization of the data before adding to the graph.